Moved column generation out to its own file and changed hidden column functionality to use the setHiddenColumns function from react table.
